Demo2/requirements.txt
numpy>=1.26.0
Pillow>=10.0.0 